pyproject.toml
[tool.poetry]
name = "mpesa-ecosystem"
version = "0.1.0"
description = ""
authors = ["Koech Hezekiah <hezekiahkoech@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-rest-framework = "^0.1.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
